added message service

+++ src/app/basket/basket.component.ts
@@ -6,10 +6,11 @@ import {
 } from '@angular/core';
 import { Location, AsyncPipe } from '@angular/common';
 import { RouterLink } from '@angular/router';
-import { BasketContainerComponent } from '../basket-container/basket-container.component';
-import { CheckboxService } from '../../services/checkbox.service';
+import { BasketContainerComponent } from './basket-container/basket-container.component';
+import { CheckboxService } from '../services/checkbox.service';
 import { map } from 'rxjs';
-import { TaskService } from '../../services/task.service';
+import { TaskService } from '../services/task.service';
+import { MessageService } from '../services/message.service';
 
 @Component({
   selector: 'app-basket',
@@ -22,6 +23,7 @@ import { TaskService } from '../../services/task.service';
 export default class BasketComponent implements OnDestroy {
   private checkboxService = inject(CheckboxService);
   private taskService = inject(TaskService);
+  private messageService = inject(MessageService);
 
   public status$ = this.checkboxService.statement$.pipe(
     map((value) => {
@@ -38,17 +40,19 @@ export default class BasketComponent implements OnDestroy {
         (<HTMLInputElement>event.target).checked,
       );
     });
-    console.log(this.checkboxService.statement$.value);
   }
-
   public restoreTasks() {
+    if (!this.checkboxService.anyChecked) return;
     this.checkboxService.trueIds.map((i) => this.taskService.restore(i));
     this.checkboxService.trueIds.map((i) => this.checkboxService.remove(i));
+    this.messageService.log('Tasks restored!', 'success');
   }
 
   public removeTasks() {
+    if (!this.checkboxService.anyChecked) return;
     this.checkboxService.trueIds.map((i) => this.taskService.remove(i));
     this.checkboxService.trueIds.map((i) => this.checkboxService.remove(i));
+    this.messageService.log('Tasks removed!', 'error');
   }
 
   public goBack() {
